FIX improving buffer reference logic

+++ src/animation/AnimationSystem.cpp
@@ -136,8 +136,7 @@ namespace animation
 
         auto& rigNodeRegistry = *m_rigNodeRegistry;
 
-        util::BufferReference rigRef = rigNodeRegistry.allocate(
-            static_cast<uint32_t>(rig.m_nodes.size()));
+        util::BufferReference rigRef = rigNodeRegistry.allocate(rig.m_nodes.size());
 
         if (rigRef.size > 0) {
             // Initialize joints
@@ -163,8 +162,7 @@ namespace animation
         auto& rigNodeRegistry = *m_rigNodeRegistry;
         auto& socketRegistry = *m_socketRegistry;
 
-        util::BufferReference socketRef = socketRegistry.allocate(
-            static_cast<uint32_t>(rig.m_sockets.size()));
+        util::BufferReference socketRef = socketRegistry.allocate(rig.m_sockets.size());
 
         // NOTE KI all joints are initially identity matrix
         // NOTE KI sockets need to be initialiazed to match initial static joint hierarchy
@@ -196,8 +194,7 @@ namespace animation
         auto& rigNodeRegistry = *m_rigNodeRegistry;
         auto& jointRegistry = *m_jointRegistry;
 
-        util::BufferReference jointRef = jointRegistry.allocate(
-            static_cast<uint32_t>(jointContainer.size()));
+        util::BufferReference jointRef = jointRegistry.allocate(jointContainer.size());
 
         // Initialize joints
         if (jointRef.size > 0) {
